+++ docs/CHANGELOG.md
@@ -5,8 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 03/3/2025
 
-## [1.0.0] - TBD
+### Added
+ - Modularized import system further using mixins
+ - Mixins are imported in task base class
+ - Pluggins added for custom import behavior
+ - Refresh files button to autoclean_review
+ - Complete documentation site
+
+### Deprecated  
+ - Most basic step functions
+
+[1.1.0]: https://github.com/cincibrainlab/autoclean_pipeline/releases/tag/v1.1.0
+
+## [1.0.0] - 02/28/2025
 
 ### Added
 - Initial release of AutoClean EEG
@@ -17,4 +30,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Database-backed processing tracking
 - Task-based modular architecture
 
-[1.0.0]: https://github.com/yourusername/autoclean-eeg/releases/tag/v1.0.0 
+[1.0.0]: https://github.com/cincibrainlab/autoclean_pipeline/releases/tag/v1.0.0